+++ .gitignore
@@ -104,4 +104,4 @@ venv.bak/
 .mypy_cache/
 
 # My ignore
-OUTPUT_TESTS
+.*
