+++ transphire_transform/dump_load/star.py
@@ -21,9 +21,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-from typing import List
+from typing import List, Tuple
 import pandas as pd # type: ignore
-from . import load_dump_util
+from . import util
 
 
 def create_star_header(names: List[str]) -> List[str]:
@@ -42,7 +42,7 @@ def create_star_header(names: List[str]) -> List[str]:
         '',
         'loop_',
         ]
-    output_list.extend(load_dump_util.create_header(names=names, index=True))
+    output_list.extend(util.create_header(names=names, index=True))
     return output_list
 
 
@@ -58,9 +58,59 @@ def dump_star(file_name: str, data: pd.DataFrame) -> None:
     None
     """
     header: List[str] = create_star_header(names=data.keys())
-    load_dump_util.dump_file(
+    util.dump_file(
         file_name=file_name,
         data=data,
         header=header,
         vertical=True
         )
+
+
+def load_star_header(file_name: str) -> Tuple[List[str], int]:
+    """
+    Load the header information.
+
+    Arguments:
+    file_name - Path to the file that contains the header.
+
+    Returns:
+    List of header names, rows that are occupied by the header.
+    """
+    start_header: bool = False
+    header_names: List[str] = []
+    idx: int
+
+    with open(file_name, 'r') as read:
+        for idx, line in enumerate(read.readlines()):
+            if line.startswith('_'):
+                if start_header:
+                    header_names.append(line.strip().split()[0])
+                else:
+                    start_header = True
+                    header_names.append(line.strip().split()[0])
+            elif start_header:
+                break
+
+    if not start_header:
+        raise IOError(f'No header information found in {file_name}')
+
+    return header_names, idx
+
+
+def load_star(file_name: str) -> pd.DataFrame:
+    """
+    Load a relion star file.
+
+    Arguments:
+    file_name - Path to the relion star file
+
+    Returns:
+    Pandas dataframe containing the star file
+    """
+    header_names: List[str]
+    skip_index: int
+    star_data: pd.DataFrame
+
+    header_names, skip_index = load_star_header(file_name=file_name)
+    star_data = util.load_file(file_name, names=header_names, skiprows=skip_index)
+    return star_data
